buttons.isWork(sql.inject(true));

+++ src/com/web/Admin/Users.java
@@ -12,10 +12,12 @@ import java.util.List;
 import static com.dao.UserDao.*;
 
 
-public class Users extends HttpServlet {
+public class Users extends HttpServlet
+{
 	private static final long serialVersionUID = 1L;
 	
-	public Users() {
+	public Users()
+	{
 		super();
 	}
 	
@@ -26,11 +28,12 @@ public class Users extends HttpServlet {
 	 * @throws IOException
 	 */
 	@Override
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		if(LoginUtil.TestAdmin(request, response))
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException
+	{
+		if (LoginUtil.TestAdmin(request, response))
 		{
 			HttpSession session = request.getSession();
-			int user_id = (int)session.getAttribute(("user_id"));
+			int user_id = (int) session.getAttribute(("user_id"));
 			boolean admin = isAdmin(user_id);
 			List<User> userList = UserDao.GetUsers();
 			
@@ -40,7 +43,7 @@ public class Users extends HttpServlet {
 			request.getRequestDispatcher("/WEB-INF/admin/users.jsp").forward(request, response);
 		}
 		
-
+		
 	}
 	
 	/**
@@ -50,24 +53,34 @@ public class Users extends HttpServlet {
 	 * @throws IOException
 	 */
 	@Override
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		String type = request.getParameter("type");
-		String value = request.getParameter("value");
-		
-		if(type.equals("username") || type.equals("password") || type.equals("progress_learn_category") || type.equals("progress_learn_lesson")) {
-			int user = Integer.parseInt(request.getParameter("user"));
-			UpdateUsers(user, type, value);
-		} else if(type.equals("delete") && !isAdmin(Integer.parseInt(request.getParameter("user")))) {
-			int user = Integer.parseInt(request.getParameter("user"));
-			DeleteUser(user);
-		} else if(type.equals("add")) {
-			String username = request.getParameter("username");
-			String password = request.getParameter("password");
-			int categoryProgress = Integer.parseInt(request.getParameter("categoryProgress"));
-			int lessonProgress = Integer.parseInt(request.getParameter("lessonProgress"));
-			int admin = Integer.parseInt(request.getParameter("admin"));
-			AddUser(username, password, categoryProgress, lessonProgress, admin);
-			response.sendRedirect("/Admin/Users");
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException
+	{
+		if (LoginUtil.TestAdmin(request, response))
+		{
+			if (LoginUtil.TestAdmin(request, response))
+			{
+				String type = request.getParameter("type");
+				String value = request.getParameter("value");
+				
+				if (type.equals("username") || type.equals("password") || type.equals("progress_learn_category") || type.equals("progress_learn_lesson"))
+				{
+					int user = Integer.parseInt(request.getParameter("user"));
+					UpdateUsers(user, type, value);
+				} else if (type.equals("delete") && !isAdmin(Integer.parseInt(request.getParameter("user"))))
+				{
+					int user = Integer.parseInt(request.getParameter("user"));
+					DeleteUser(user);
+				} else if (type.equals("add"))
+				{
+					String username = request.getParameter("username");
+					String password = request.getParameter("password");
+					int categoryProgress = Integer.parseInt(request.getParameter("categoryProgress"));
+					int lessonProgress = Integer.parseInt(request.getParameter("lessonProgress"));
+					int admin = Integer.parseInt(request.getParameter("admin"));
+					AddUser(username, password, categoryProgress, lessonProgress, admin);
+					response.sendRedirect("/Admin/Users");
+				}
+			}
 		}
 	}
 }
